usrspc: change RHand shift mod taps back to right shift
70e9618698a87f3f0713d5c27aa42f5a43a5dddf

In the commit above, I changed K to be LShift in an attempt to avoid an
auto-shift related issue. The issue is specific to combo keys, and looks
like it's removing the RShift flag for the combo's key up event,
preventing the host OS from considering the key as shifted.

Changing K to RShift brings back this issue, but I plan to disable auto
shift; having Shift next to the LWR mod tap key is more useful.

Having K as RShift avoids a different issue:

If QKC_K is held while QKC_D is already held, after releasing only K,
MOD_LSFT is disabled. It's an issue in lazygit when I mistakenly hold K
for a little too long when scrolling a diff (which requires holding
shift).